backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
SQLAlchemy
psycopg2-binary
python-dotenv
celery
redis
gunicorn